requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Document Generation
reportlab>=4.0.0
PyPDF2>=3.0.0

# Security & Authentication
bcrypt>=4.0.0
PyJWT>=2.8.0

# Database
sqlalchemy>=2.0.0

# Additional Utilities
python-dateutil>=2.8.0
pytz>=2023.3.post1
Pillow>=10.0.0

# Development & Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
